feat: enhance organization permissions handling and add user permission checks

+++ modules/services/api/src/org/controllers/organizations.controller.ts
@@ -4,7 +4,9 @@ import {
   Body,
   Controller,
   Delete,
+  ForbiddenException,
   Get,
+  NotFoundException,
   Param,
   ParseUUIDPipe,
   Patch,
@@ -15,6 +17,8 @@ import {
   ApiOkResponse,
   ApiOperation,
 } from '@nestjs/swagger';
+import { UserWithPermissions } from '@vidya/api/auth/decorators';
+import { UserPermissions } from '@vidya/api/auth/utils';
 import * as dto from '@vidya/api/org/dto';
 import * as entities from '@vidya/entities';
 import { Routes } from '@vidya/protocol';
@@ -30,6 +34,10 @@ export class OrganizationsController {
     @InjectMapper() private readonly mapper: Mapper,
   ) {}
 
+  /* -------------------------------------------------------------------------- */
+  /*                                  GET /orgs                                 */
+  /* -------------------------------------------------------------------------- */
+
   @Get(Routes().org.find())
   @ApiOperation({
     summary: 'Returns a list of organizations',
@@ -39,8 +47,13 @@ export class OrganizationsController {
     type: GetOrganizationsResponse,
     description: 'Get a list of organizations.',
   })
-  async getOrganizations(): Promise<GetOrganizationsResponse> {
-    const orgs = await this.organizationsService.findAllBy({});
+  async getOrganizations(
+    @UserWithPermissions() userPermissions: UserPermissions,
+  ): Promise<GetOrganizationsResponse> {
+    const orgs =
+      await this.organizationsService.findPermittedOrganizationsBy(
+        userPermissions,
+      );
     return {
       items: this.mapper.mapArray(
         orgs,
@@ -50,18 +63,40 @@ export class OrganizationsController {
     };
   }
 
+  /* -------------------------------------------------------------------------- */
+  /*                                GET /orgs/:id                               */
+  /* -------------------------------------------------------------------------- */
+
   @Get(Routes().org.get(':id'))
   async getOrganization(
     @Param('id', new ParseUUIDPipe()) id: string,
+    @UserWithPermissions() userPermissions: UserPermissions,
   ): Promise<dto.GetOrganizationResponse> {
-    const org = await this.organizationsService.findOneBy({ id });
-    return this.mapper.map(org, entities.Organization, dto.Organization);
+    const orgs = await this.organizationsService.findPermittedOrganizationsBy(
+      userPermissions,
+      { id },
+    );
+    if (orgs.length === 0) {
+      throw new NotFoundException(`Organization with id ${id} not found`);
+    }
+    return this.mapper.map(orgs[0], entities.Organization, dto.Organization);
   }
+
+  /* -------------------------------------------------------------------------- */
+  /*                                 POST /orgs                                 */
+  /* -------------------------------------------------------------------------- */
 
   @Post(Routes().org.create())
   async createOrganization(
     @Body() request: dto.CreateOrganizationRequest,
+    @UserWithPermissions() userPermissions: UserPermissions,
   ): Promise<dto.CreateOrganizationResponse> {
+    if (!userPermissions.hasPermissions(['orgs:create'])) {
+      throw new ForbiddenException(
+        'User does not have permission to create organizations',
+      );
+    }
+
     const entity = await this.organizationsService.create({
       name: request.name,
     });
@@ -72,11 +107,17 @@ export class OrganizationsController {
     );
   }
 
+  /* -------------------------------------------------------------------------- */
+  /*                               PATCH /orgs/:id                              */
+  /* -------------------------------------------------------------------------- */
+
   @Patch(Routes().org.update(':id'))
   async updateOrganization(
     @Body() request: dto.UpdateOrganizationRequest,
     @Param('id', new ParseUUIDPipe(), OrganizationExistsPipe) id: string,
+    @UserWithPermissions() userPermissions: UserPermissions,
   ): Promise<dto.UpdateOrganizationResponse> {
+    this.checkUserPermissions(userPermissions, ['orgs:update'], id);
     const org = await this.organizationsService.updateOneBy({ id }, request);
     return this.mapper.map(
       org,
@@ -85,14 +126,38 @@ export class OrganizationsController {
     );
   }
 
+  /* -------------------------------------------------------------------------- */
+  /*                              DELETE /orgs/:id                              */
+  /* -------------------------------------------------------------------------- */
+
   @Delete(Routes().org.delete(':id'))
   @ApiNotFoundResponse({
     description: 'Organization not found',
   })
   async deleteOrganization(
     @Param('id', new ParseUUIDPipe(), OrganizationExistsPipe) id: string,
+    @UserWithPermissions() userPermissions: UserPermissions,
   ): Promise<dto.DeleteOrganizationResponse> {
+    this.checkUserPermissions(userPermissions, ['orgs:delete'], id);
     await this.organizationsService.deleteOneBy({ id });
     return { success: true };
+  }
+
+  /* -------------------------------------------------------------------------- */
+  /*                               Private methods                              */
+  /* -------------------------------------------------------------------------- */
+
+  private checkUserPermissions(
+    userPermissions: UserPermissions,
+    requiredPermissions: string[],
+    id: string,
+  ) {
+    if (
+      !userPermissions
+        .getPermittedOrganizations(requiredPermissions)
+        .includes(id)
+    ) {
+      throw new ForbiddenException('User does not have permission');
+    }
   }
 }
